Flyttar ansvaret for koplinga sakstype <-> behandlingsnummer eitt steg i rett retning
